Wrote components to remove redundant code, Changes the layout of the header, Fixed mobile pagination buttons blocked.
