+++ src/App.tsx
@@ -7,23 +7,26 @@ import Login from "./sub-pages/Login";
 import SignUp from "./sub-pages/SignUp";
 import Settings from "./sub-pages/Settings";
 import { SearchProvider } from "./Contexts/SearchContext";
+import { UserProvider } from "./Contexts/UserContext";
 
 function App() {
-return (
-<>
-<Header />
-<SearchProvider>
-<Routes>
-<Route path="/" element={<Search />} />
-<Route path="/search" element={<Search />} />
-<Route path="/wish-list" element={<WishList />} />
-<Route path="/what-is-verifind" element={<WhatIsVerifind />} />
-<Route path="/login" element={<Login />} />
-<Route path="/signup" element={<SignUp />} />
-<Route path="/settings" element={<Settings />} />
-</Routes>
-</SearchProvider>
-</>
+  return (
+    <>
+      <Header />
+      <SearchProvider>
+        <UserProvider>
+          <Routes>
+            <Route path="/" element={<Search />} />
+            <Route path="/search" element={<Search />} />
+            <Route path="/wish-list" element={<WishList />} />
+            <Route path="/what-is-verifind" element={<WhatIsVerifind />} />
+            <Route path="/login" element={<Login />} />
+            <Route path="/signup" element={<SignUp />} />
+            <Route path="/settings" element={<Settings />} />
+          </Routes>
+        </UserProvider>
+      </SearchProvider>
+    </>
   );
 }
