+++ src/Contexts/UserContext.tsx
@@ -1,0 +1,41 @@
+// src/contexts/UserContext.tsx
+import { createContext, useContext, useEffect, useState } from "react";
+import type { ReactNode } from "react";
+import { supabase } from "../supabase-client";
+
+type UserContextType = {
+  username: string | null;
+  setUsername: React.Dispatch<React.SetStateAction<string | null>>;
+  userId: string | null;
+  setUserId: React.Dispatch<React.SetStateAction<string | null>>;
+};
+
+const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export const UserProvider = ({ children }: { children: ReactNode }) => {
+  const [username, setUsername] = useState<string | null>(null);
+  const [userId, setUserId] = useState<string | null>(null);
+
+  useEffect(() => {
+    supabase.auth.getSession().then(({ data }) => {
+      const name = data.session?.user?.user_metadata?.username ?? null;
+      const id = data.session?.user?.id ?? null;
+      setUsername(name);
+      setUserId(id);
+    });
+  }, []);
+
+  return (
+    <UserContext.Provider value={{ username, userId, setUsername, setUserId }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
+
+export const useUser = (): UserContextType => {
+  const context = useContext(UserContext);
+  if (!context) {
+    throw new Error("useUser must be used within a UserProvider");
+  }
+  return context;
+};
